% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arakno.R
\docType{data}
\name{globalTree}
\alias{globalTree}
\title{Global spider backbone tree.}
\format{
A phylo object with 132 families and 800+ genera, 1400+ species.
}
\usage{
data(globalTree)
}
\description{
From Macias-Hernandez et al. (2020) with nomenclature updated.
}
\keyword{datasets}
